+++ src/app/[locale]/error/page.tsx
@@ -1,7 +1,7 @@
 import { Route } from "next";
 import { StatusCodeView } from "@/components/common/display/status/status-code-view";
 import { Button } from "@/components/common/ui/button";
-import { Link } from "@/i18n/routing";
+import { Link } from "@/infrastructures/i18n/routing";
 
 // Auth0 Error page
