+++ src/infrastructures/i18n/request.ts
@@ -12,6 +12,6 @@ export default getRequestConfig(async ({ requestLocale }) => {
 
 	return {
 		locale,
-		messages: (await import(`../../messages/${locale}.json`)).default,
+		messages: (await import(`../../../messages/${locale}.json`)).default,
 	};
 });
