+++ src/middleware.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import createMiddleware from "next-intl/middleware";
 import { auth } from "@/infrastructures/auth/auth-provider";
-import { routing } from "./i18n/routing";
+import { routing } from "./infrastructures/i18n/routing";
 
 const handleI18nRouting = createMiddleware(routing);
